chore: get rid of temp page, add to home, cleanup props
