+++ .env.example
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_SUBGRAPH_URL=https://api.thegraph.com/subgraphs/name/nounsdao/nouns-subgraph
