+++ hooks/useIsMounted.ts
@@ -1,5 +1,6 @@
 import * as React from 'react';
 
+//https://github.com/wagmi-dev/wagmi/issues/542
 export const useIsMounted = () => {
   const [mounted, setMounted] = React.useState(false);
