--- pages/api/proposals/index.ts
@@ -1,20 +1,0 @@
-import { NextApiRequest, NextApiResponse } from 'next';
-import { subgraphService } from '../../../lib/services/subgraph.service';
-import { restrictHandlerMethods } from '../../../lib/util/api';
-
-async function handler(req: NextApiRequest, res: NextApiResponse) {
-  try {
-    // TODO - should service default
-    const proposals = await subgraphService.getOpenProposals('desc', 10, 0);
-    res.setHeader('Cache-Control', 'no-cache');
-    res.status(200).json(proposals);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({
-      statusCode: 500,
-      message: 'An error occurred while fetching proposals, please try again.',
-    });
-  }
-}
-
-export default restrictHandlerMethods(handler, ['GET']);
